Rewrite peer-stream stuff
