+++ zeronet-client/package.json
@@ -24,7 +24,7 @@
   "dependencies": {
     "bl": "^1.2.1",
     "clone": "^2.1.1",
-    "data-queue": "0.0.2",
+    "data-queue": "0.0.3",
     "msgpack5": "^3.6.0",
     "pull-queue": "^0.1.4",
     "pull-stream": "^3.6.1",
