+++ zeronet-common/package.json
@@ -25,7 +25,7 @@
     "base-x": "^3.0.2",
     "bittorrent-tracker": "^9.2.4",
     "bunyan": "^1.8.12",
-    "data-queue": "0.0.2",
+    "data-queue": "0.0.3",
     "hypercache": "^0.2.0",
     "ip": "^1.1.5",
     "merge-recursive": "0.0.3",
