+++ zeronet-crypto/package.json
@@ -25,7 +25,7 @@
   "homepage": "https://github.com/ZeroNetJS/zeronet-js#readme",
   "dependencies": {
     "bitcoinjs-message": "^2.0.0",
-    "data-queue": "0.0.2",
+    "data-queue": "0.0.3",
     "interface-connection": "^0.3.2",
     "libp2p-secio": "^0.8.1",
     "node-forge": "^0.7.1",
